+++ src/appleseed/renderer/kernel/shading/shadingpoint.cpp
@@ -856,8 +856,7 @@ void ShadingPoint::initialize_osl_shader_globals(
         m_shader_globals.N = Vector3f(get_original_shading_normal());
         m_shader_globals.Ng = Vector3f(get_geometric_normal());
 
-        m_shader_globals.backfacing =
-            get_side() == ObjectInstance::FrontSide ? 0 : 1;
+        m_shader_globals.backfacing = get_side() == ObjectInstance::FrontSide ? 0 : 1;
 
         // Surface parameters and their differentials.
         const Vector2f& uv = get_uv(0);
@@ -889,8 +888,8 @@ void ShadingPoint::initialize_osl_shader_globals(
         m_shader_globals.dPdv = Vector3f(get_dpdv(0));
 
         // Time and its derivative.
-        m_shader_globals.time = static_cast<float>(ray.m_time.m_absolute);
-        m_shader_globals.dtime = static_cast<float>(m_scene->get_active_camera()->get_shutter_open_time_interval());
+        m_shader_globals.time = ray.m_time.m_absolute;
+        m_shader_globals.dtime = m_scene->get_active_camera()->get_shutter_open_time_interval();
 
         // Velocity vector.
         m_shader_globals.dPdtime =
@@ -927,7 +926,7 @@ void ShadingPoint::initialize_osl_shader_globals(
     // Always update the surface area of emissive objects.
     m_shader_globals.surfacearea =
         ray_flags == VisibilityFlags::LightRay && sg.has_emission()
-            ? static_cast<float>(sg.get_surface_area(&get_assembly_instance(), &get_object_instance()))
+            ? sg.get_surface_area(&get_assembly_instance(), &get_object_instance())
             : 0.0f;
 
     // Output closure.
